server/requirements.txt
numpy==1.25.2
pandas==1.5.3
pandas-datareader==0.10.0
pandas-gbq==0.19.2
pandas-stubs==1.5.3.230304
Pillow==9.4.0
pip-tools==6.13.0
requests==2.31.0
requirements-parser==0.5.0
scikit-image==0.19.3
scikit-learn==1.2.2
scipy==1.11.4
sklearn-pandas==2.2.0
